requirements.txt
amqp==5.2.0
ansicon==1.89.0
anyio==4.2.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
blessed==1.20.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
flake8==6.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
inflection==0.5.1
jinxed==1.2.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
mccabe==0.7.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
referencing==0.32.0
rpds-py==0.15.2
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
